Added opcua-version that provides functional authentication via certificate and sign/encrypt. Added example certificate for login. Modified Dockerfile for adding certificates into container.
