+++ software/opcua/opcua_org.py
@@ -25,12 +25,12 @@ async def main():
 
     # populating the cybics address space
     # server.nodes, contains links to very common nodes like objects and root
-    myobj = await server.nodes.objects.add_object(idx, "MyObject_TEST")
+    myobj = await server.nodes.objects.add_object(idx, "MyObject")
     gstvar = await myobj.add_variable(idx, "GST", ua.UInt16(0))
     hptvar = await myobj.add_variable(idx, "HPT", ua.UInt16(0))
     systemSenvar = await myobj.add_variable(idx, "systemSen", ua.UInt16(0))
     boSenvar = await myobj.add_variable(idx, "boSen", ua.UInt16(0))
-    stopvar = await myobj.add_variable(idx, "STOP", ua.UInt16(0))
+    stopvar = await myobj.add_variable(idx, "stop", ua.UInt16(0))
     manualvar = await myobj.add_variable(idx, "manual", ua.UInt16(0))
     await server.nodes.objects.add_method(
         ua.NodeId("ServerMethod", idx),
