+++ training/opcua_ctf/peer-private-key-example-1.pem
@@ -1,0 +1,28 @@
+-----BEGIN PRIVATE KEY-----
+MIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDAlV4LpXajLq54
+D+CJrvkQWB/SmjHrY1uIvcS+6H3UoDhdRjcsys/Pp9aqXuj/MBwhKOwgeLW+H61f
+pulOOQ/23PADqwGUE5WzGtBg7/HnXQrgdETLdW8kta4H/rbTZYNmvzfXoWW6o8JA
+68ES523RDvTUVApF1tN1BwJ9kCnVHD5kTZTrz560eRIBeo8Gy3vGDsGQ7ZSvdVu7
+3hTS9DYpDEAmVofRXyu6H+6xjRMTqrE1F/h18/7XoKh6CaNAQAj5LwFEoQsVIH0F
+0CEqJz/TQH48Nx6cDBzKLyA688CPgkyP/rDm7O+/lhK8PvyaQMhkExFgiAS+PmC+
+xwvbWUeFAgMBAAECggEBAL9dLSBhgKvSHSfMjxABQBhPmP+KnQQLrM8J0tDo8PAx
+bRCyTu1BcpIjcxnESb4Zrjn9EfWKPP35OCwODVLfPM/O5uieKmlbtKtOf7VzeGqw
+d7hlkF6kCJbvuUjUgBM7kgCtDe5EwHNfeg1MtoP0V5kfRYUfJT4Uo/Zryru+XvMU
+7ibCc+4ME14u75CgfvI0YPC9vxQVM2s7vZOl0u0pBz7p6FGViXQdzeyDEmFEfaNJ
+6hFtrkUgr+eYvPBOtHgPWMmos4xeMW7OsNGoMCXqLJkrkqWEv2mlQk05FMEnQHzs
+u+GsSiPufm5PCKBU+8k86iC/0fb/6NL14A1w2b3aR1kCgYEA5QBcqHsQP3TQt1uH
+6KuaSjehJ4ubyWaSYh5i9oTatBjDq11nm3yizJsTrAlq/m1wvlnyvF5pajW3tgpN
+cE3rzB6COzls/lT5WWqXU8+NVKlbNDoKgxuVbH4PFBivE7yNSi7I6KyaLyzrSAls
+qSIm6h6aAYDkqKnORls1lT8TX9MCgYEA10naIOMQPJtgYH/Y+UKQCcsA4F9C985z
+0Qp5++V4WQf8AbebC6W6F2S79eer9roOdy3LJLLm3tmmcluP7pUBgDSWsSBOb58L
+eT1cAHkrB1IWHSaKBaO3BmTgsbPKAMyVexYB/qoi8OQA3PT7tkes2pvi3EFXfY+2
+w6X59GkC/EcCgYBs/twbUN+t5FLHQuiQRMUeGBY7NUbD+G8qBl6CCT/zQmo3cTC0
+IAF4s0MMujrTuJFV+IgBQkqYQ9MZ8RNkpaHauJaxy6Zl81nvRj2apYoWRrljRVDl
+kEZxqhpRLWJ/WVpCebwkiqqkU5P06I9wfRHakxAoMQokgRXuBNGfJLN7AwKBgGJi
+r7qjygHIiAzRoFiZILY4/ZeMMVCKGseHYCPLxGn7iiFGxSZXgGm8cWF5lE7WzFGO
+re2BjJx/Ht/YFDjWA6+RWORtfSrpphasAeV9ejqIOu7rc9wL61KSu1S6czBWTmHH
+tvWozvsDV5mio32z17y3zL8XZWJOURmsOkSF0QV1AoGAd/eXssJE/cwq+oJsjc1S
+4K/tbx+eq8luc+o1hInOGTicJbuPKO3ggJZw0TSx+pA0ZH1ktCzLDSY45e8Xrogj
+umNTJw8pg7OlKhdxpJfvWnQ4LNCwIjQuhHJ2SQcXgyIyH7dzLT1kQJ96+q1zlQAH
+d4f9Cg9XV95DytBuvJaxrNM=
+-----END PRIVATE KEY-----
