+++ training/opcua_ctf/test_credentials.txt
@@ -1,0 +1,6 @@
+admin 1234567
+admin admin
+user1 pw1
+Admin Admin
+ 
+blub a
